requirements.txt
alembic==1.4.3
click==7.1.2
Flask==1.1.2
Flask-Admin==1.5.7
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-SocketIO==5.0.1
Flask-Cors==3.0.10
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
PyJWT==1.7.1
six==1.15.0
SQLAlchemy==1.3.21
Werkzeug==1.0.1
